requirements.txt
kafka-python
numpy
Pillow